add : return id when add table

+++ dss-apps/dss-datamodel-center/dss-datamodel-table-management/src/main/java/com/webank/wedatasphere/dss/datamodel/table/service/TableService.java
@@ -15,7 +15,7 @@ public interface TableService extends IService<DssDatamodelTable> {
      * @param vo
      * @return
      */
-    int addTable(TableAddVO vo)  throws ErrorException;
+    Long addTable(TableAddVO vo)  throws ErrorException;
 
 
     /**
